Move pytest_assertrepr_compare to python_api.py
